requirements.txt
asgiref==3.3.1
autopep8==1.5.4
dj-database-url==0.5.0
dj-rest-auth==2.1.3
Django==3.1.5
django-cors-headers==3.7.0
django-heroku==0.3.1
django-nested-admin==3.3.3
djangorestframework==3.12.2
gunicorn==20.0.4
psycopg2==2.8.6
pycodestyle==2.6.0
python-decouple==3.4
python-monkey-business==1.0.0
pytz==2020.5
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
whitenoise==5.2.0
